Add routes for obtaining proxy guids
Added routes for obtaining the proxy guids by either the proxy id
or by the tool consumer id, which is the same as the assignment id in
Canvas. This should allow us to use the api to tell the test tool to go
fetch a submission that it received through the live events flow, which
we can then tell the tool to submit an originality report for.