Update docker-compose and fix travis.yml
The docker-compose didn't specify a postgres password, which meant that
it wouldn't work at all unless a developer modified it themselves or
created an override file. This commit fixes that.

In addition, Travis was running all tests inside of its VM, instead of
in the Docker container. This was causing tool versioning issues.
Luckily, Docker is perfectly suited to solve this problem, so we'll just
run all of our tests inside of the web Docker container now, as it has
the correct version of Ruby, Bundle, node, etc, and we don't have to
worry about configuring Travis in addition to Docker.