Fix CI and docker-compose
Previously, our CI was trying to run all tests inside of a VM, instead
of using Docker/docker-compose. This led to versioning issues with
tooling, so switching to docker for running all of our tests should help
immensely, in addition to reducing the amount of configuration we have
to do.

In addition, the docker-compose file needed to be fixed, as it didn't
specify a password for Postgres, which meant it was effectively useless.
That should be fixed now.