Add routes for checking status
Added two routes for checking whether a given assignment or submission
exists in the test tool. The primary purpose of this is for integration
testing, as all the endpoints use Canvas specific information, like
lti_submission_id and tc_id, to actually fetch the corresponding
assignment/submission, rather than the actual id of the assignment.